requirements.txt
django
djangorestframework
dj_rest_auth
pandas